finishedFunc called whenever needed
until some weeks ago, the writeFunctionStack in roomServer actually did not work, because this.busy was always false; now, this.busy is used, but in some cases (e.g. on errors), this,.busy was not reset to false, or the next write funciton was not called respectively. this is hopefully solved now!