findRoom await room ready
Provide the possibility (by default true) to wait for a room to get ready before it is returned by the findRoom function. This may help a lot, since currently, if a client requests a subroom that does not exist, the room will get prepared, but the client then fails to enter the room since the room is not ready yet during this request. Only when the client retries after 1s, it is likely that now it can enter the room. Now, the waiting time is not n*1s, but only as long as needed.