roomManager / roomClient separating data for room and data for roomManager
So far, the roomManager-vue instance contained every room, including its data. (And owed to the reference to the roomManager, also the roomManager itself was proxied by the roomManager-vue.) That means, both the regular vue as well as the roomManager-vue instances proxy the same data. This worked mostly so far, except when a fullReload had to be done. Then the proxy of the room got broken and room.data != vue.data, even directly after setprops.
The double-proxy of the data was obviously the problem. Now, the roomManager-vue only gets the necessary information, without the fullroom and the room.data. roomClient had to be changed t notify the roomManager about the relevant changes for roomManager-vue.
Additionally, avoid the double call to dataArrived/fullReload at the first connection.