do NOT use findPyPK in updateXYZ
findByPk will return a new sequelize model, which would get updated there after. However, we MUST use the original model object to ensure that the references to subtables are correctly kept. (Example: the eventGroup should contain eventGroup.rounds, which is ensured by having the proper includes in the select-call (findAll or similar) to sequelize AND (for update and create) having "rounds" in the input data. By using findByPk without the include, the "rounds" reference is lost and since the data in the update on purpose does not have "rounds" in it, the updated model would not have the properties any more*. (* the updated model is the one findByPk, which never had the rounds property)