fix error due to multi-used sid
do NOT use the sid anymore to reference the online/offlineWritingClients, but use the tabId; by using the sid, we create a crash when two tabs exist int he same browser and read the same data; when the first client/tabId left, this obviously removed also the other clients from the clients list and resulted in a crash when the second client/tabId left.
fixed by using the tabId for as the reference in online/offlineWritingClients. Also changed to revokeWritingTicket by the hashed tabId instead of the hashed sid.