replace contest watcher
Before, changes to contest i vue were (nearly) directly done on the main data  and we had a contest watcher to propagate any changes to the server. This complex construct did not work properly. Now, the changes on the actual data are only done in rContestClient and not in the vue anymore.
This should also fix the problem that sometimes the change to the status was not propagated.